src/requirements.txt
docker
requests
pyyaml